Add rewards to move popup
