+++ src/app/d2-vendors/vendor-item-dialog.html
@@ -2,6 +2,6 @@
      item-popup
      item-popup-boundary-class="store-bounds"
      dim-click-anywhere-but-here="closeThisDialog()">
-  <dim-move-item-properties item="vm.item" failure-strings="vm.vendorItem.failureStrings"></dim-move-item-properties>
+  <dim-move-item-properties item="vm.item" failure-strings="vm.vendorItem.failureStrings" rewards="vm.rewards"></dim-move-item-properties>
   <div class="arrow" ng-class="'is-{{vm.item.dmg}}'"></div>
 </div>
